HyperV: Fix vm disk path issue
As the disk number of iSCSI attached disks can change after host
reboot, passthrough attached volumes can get attached in this case.

This bug was partially fixed during Icehouse by this patch:
https://review.openstack.org/95356

One of the issues with this patch is that it only handles SCSI
attached disks, for which reason this issue continues to occur when
having generation 1 VMs booted from volume, in which case the disk
will be placed on the IDE controller.

In this case, one instance may end up booting from another tenant's
volume, which is a critical security issue.

Also, it assumes that the block device info volume order matches
the according disk controller slot order, which is wrong.

This patch fixes this issue, looping through the block device
info and fixing the disk path for all the passthrough attached
disks, regardless of the disk controller.

This requires the virtual disk resources attached to the instance
to be tagged with the serial number.

Note that some unit tests have been updated to match the ones from
Nova.

Co-Authored-By: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>

(cherry-picked from commit 9de0bec4f9cdf0cd6424aeb95daabe657d4db2d4)
Closes-Bug: #1526831
Depends-On: Ibe463ce9ffb9129cab4b99fe7967ccb5f2181958
Change-Id: Icf8697d917c814ec120fce9898661eb9a20ade48