Avoid handling Hyper-V instance event objects by compute-hyperv
At the moment, the instance event handler retrieves an event
listener from os-win and polls for events, handling Hyper-V specific
instance objects, as well as WMI exceptions.

One of the goals that os-win tries to achieve is hiding Hyper-V
specific details such as those mentioned above.

This patch slightly changes the way events are handled, meaning
that we'll let os-win do the polling, which will invoke a callback
specifying the instance name and the new power state when an instance
power state change occurs.

Also, the base test class has been updated not to mock the wmi lib
as we don't directly use it anymore.

Change-Id: I5cdfb5f8ab2ca4b66cc4db56056df4537551c9bb
Depends-On: Iddc52878bcc84bf3d8c31edbcd178eac4b59092d