Avoid starting serial console workers while deleting vm files
If an instance destroy request is received really quick after it was
powered on, this can fail, having the console log files in use.

This can happen if the instance power on event is received after the
destroy operation has started, which will spawn serial console workers
after the destroy operation tries to clean them up.

We can avoid this situation by simply ensuring that while we delete
the instance log files, the serial console workers are stopped and
no other workers will be spawned.

Change-Id: I3759b557d735cde4391f94819a32994dfb184702
Closes-Bug: #1556189