SMBFS: properly handle empty connection info options
We use regex to fetch the crentials from the connection info options
when mounting SMB shares.

The issue is that this field may be empty, in which case we may get
a TypeError. This patch fixes this issue.

Change-Id: I44f8a74ad38e4419f58bf8f59682af79fff8070c
Closes-Bug: #1564829
(cherry picked from commit 01eec24a13d2886440d2ea71eb84d9598b09380c)