Hyper-V: properly handle loopback shares
At the moment, the Hyper-V driver uses the UNC path of images
stored on SMB shares, regardless if the share is remote or not.
According to the MS docs, this is not supported for SOFS SMB shares.

This is troublesome for the Hyper-C scenario, as Hyper-V will attempt
to modify the image ACLs, making them unusable.

The fix consists in checking if the share is local, and use the
local path in that case.

Closes-Bug: #1580122

(cherry-picked from b964f2bce3f3eeac3659e5ef50f520ee1da4afb0)
Change-Id: Ib736b39f3bfa58b2737647de23f741ffcf85180b